added image ordering

+++ server/src/application/use-cases/image/deleteImageUseCase.ts
@@ -1,8 +1,21 @@
+import path from "path";
+import fs from 'fs/promises'
 import { IImageRepository } from "../../interfaces/IImageRepository";
 
 export class DeleteImageUseCase {
     constructor(private imageRepository: IImageRepository){}
     async execute(id: string) :Promise<void>{
-        await this.imageRepository.deleteImage(id)
+        try {
+            const imageData = await this.imageRepository.getImageById(id)
+        if(!imageData){
+            throw new Error('Image not found')
+        }
+        const imagePath = path.join(__dirname, '../../../../', imageData.url);
+        await fs.unlink(imagePath)
+        await this.imageRepository.deleteImage(id) 
+        } catch (error) {
+            throw error
+        }
+       
     }
-}
+}
